+++ R/fslice.R
@@ -89,21 +89,21 @@
 fslice <- function(data, ..., .by = NULL,
                    keep_order = FALSE, sort_groups = TRUE){
   dots <- list(...)
-  N <- df_nrow(data)
   n <- unlist(dots, recursive = TRUE, use.names = FALSE)
-  if (length(n) == 0L) n <- 0L
-  n_rng <- collapse::frange(n)
-  sum_n_rng <- sum(n_rng)
-  if (abs(sum_n_rng) != sum(abs(n_rng))){
+  N <- df_nrow(data)
+  if (length(n) == 0L){
+    n <- 0L
+  }
+  rng <- collapse::frange(n, na.rm = FALSE)
+  rng_sum <- sum(sign(1 / rng))
+  if (abs(rng_sum) != 2){
     stop("Can't mix negative and positive locations")
   }
-  # range_sign <- sign(sum(sign(1/n_rng))) # This can deal with -0
-  range_sign <- sign(sum_n_rng)
-  n <- as.integer(n)
+  slice_sign <- sign(rng_sum)
   # Groups
   group_vars <- get_groups(data, .by = {{ .by }})
   if (length(group_vars) == 0L){
-    if (any(abs(n_rng) > N)){
+    if (any(abs(rng) > N)){
       i <- n[cpp_which(data.table::between(n, -N, N))]
     } else {
       i <- n
@@ -119,7 +119,7 @@ fslice <- function(data, ..., .by = NULL,
     n <- n[cpp_which(data.table::between(n, -GN, GN))]
     rows <- group_df[[".loc"]]
     row_lens <- group_df[[".size"]]
-    if (range_sign >= 1){
+    if (slice_sign >= 1){
       size <- pmin.int(max(n), row_lens)
     } else {
       size <- pmax.int(0L, row_lens - max(abs(n)))
@@ -130,7 +130,7 @@ fslice <- function(data, ..., .by = NULL,
       row_lens <- row_lens[keep]
       size <- size[keep]
     }
-    if (length(n) == 1){
+    if (length(n) == 1 && slice_sign >= 1){
       i <- list_subset(rows, n)
       i <- i[cpp_which(is.na(i), invert = TRUE)]
     } else {