+++ src/cpp11.cpp
@@ -202,6 +202,13 @@ extern "C" SEXP _timeplyr_cpp_df_group_indices(SEXP rows, SEXP size) {
   END_CPP11
 }
 // utils.cpp
+SEXP cpp_r_obj_address(SEXP x);
+extern "C" SEXP _timeplyr_cpp_r_obj_address(SEXP x) {
+  BEGIN_CPP11
+    return cpp11::as_sexp(cpp_r_obj_address(cpp11::as_cpp<cpp11::decay_t<SEXP>>(x)));
+  END_CPP11
+}
+// utils.cpp
 bool cpp_any_address_changed(SEXP x, SEXP y);
 extern "C" SEXP _timeplyr_cpp_any_address_changed(SEXP x, SEXP y) {
   BEGIN_CPP11
@@ -230,10 +237,17 @@ extern "C" SEXP _timeplyr_cpp_bin_grouped(SEXP x, SEXP y, SEXP codes, SEXP right
   END_CPP11
 }
 // utils.cpp
-SEXP cpp_list_subset(SEXP x, SEXP ptype, SEXP i, SEXP replace);
-extern "C" SEXP _timeplyr_cpp_list_subset(SEXP x, SEXP ptype, SEXP i, SEXP replace) {
+SEXP cpp_list_subset(SEXP x, SEXP ptype, SEXP i, SEXP default_value);
+extern "C" SEXP _timeplyr_cpp_list_subset(SEXP x, SEXP ptype, SEXP i, SEXP default_value) {
   BEGIN_CPP11
-    return cpp11::as_sexp(cpp_list_subset(cpp11::as_cpp<cpp11::decay_t<SEXP>>(x), cpp11::as_cpp<cpp11::decay_t<SEXP>>(ptype), cpp11::as_cpp<cpp11::decay_t<SEXP>>(i), cpp11::as_cpp<cpp11::decay_t<SEXP>>(replace)));
+    return cpp11::as_sexp(cpp_list_subset(cpp11::as_cpp<cpp11::decay_t<SEXP>>(x), cpp11::as_cpp<cpp11::decay_t<SEXP>>(ptype), cpp11::as_cpp<cpp11::decay_t<SEXP>>(i), cpp11::as_cpp<cpp11::decay_t<SEXP>>(default_value)));
+  END_CPP11
+}
+// utils.cpp
+SEXP cpp_new_list(R_xlen_t size, SEXP default_value);
+extern "C" SEXP _timeplyr_cpp_new_list(SEXP size, SEXP default_value) {
+  BEGIN_CPP11
+    return cpp11::as_sexp(cpp_new_list(cpp11::as_cpp<cpp11::decay_t<R_xlen_t>>(size), cpp11::as_cpp<cpp11::decay_t<SEXP>>(default_value)));
   END_CPP11
 }
 // which.cpp
@@ -264,7 +278,9 @@ static const R_CallMethodDef CallEntries[] = {
     {"_timeplyr_cpp_lengths",              (DL_FUNC) &_timeplyr_cpp_lengths,              1},
     {"_timeplyr_cpp_list_subset",          (DL_FUNC) &_timeplyr_cpp_list_subset,          4},
     {"_timeplyr_cpp_list_which_not_null",  (DL_FUNC) &_timeplyr_cpp_list_which_not_null,  1},
+    {"_timeplyr_cpp_new_list",             (DL_FUNC) &_timeplyr_cpp_new_list,             2},
     {"_timeplyr_cpp_num_na",               (DL_FUNC) &_timeplyr_cpp_num_na,               1},
+    {"_timeplyr_cpp_r_obj_address",        (DL_FUNC) &_timeplyr_cpp_r_obj_address,        1},
     {"_timeplyr_cpp_roll_diff",            (DL_FUNC) &_timeplyr_cpp_roll_diff,            3},
     {"_timeplyr_cpp_roll_diff_grouped",    (DL_FUNC) &_timeplyr_cpp_roll_diff_grouped,    5},
     {"_timeplyr_cpp_roll_lag",             (DL_FUNC) &_timeplyr_cpp_roll_lag,             3},